requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.24.4
Pillow==10.4.0
tqdm==4.67.1



torch
torchvision
tqdm
Pillow
numpy
scipy
transformers
ftfy
regex
pycocotools
